+++ lightning/subscribe_to_open_requests.js
@@ -17,6 +17,8 @@ const weightPerVByte = 4;
   To return to default behavior of accepting all channel requests, remove all
   listeners to `channel_request`
 
+  LND 0.11.1 and below do not support `accept` or `reject` arguments
+
   {
     lnd: <Authenticated LND API Object>
   }
@@ -29,7 +31,15 @@ const weightPerVByte = 4;
 
   @event 'channel_request'
   {
-    accept: <Accept Request Function>
+    accept: <Accept Request Function> ({
+      [cooperative_close_address]: <Restrict Coop Close To Address String>
+      [min_confirmations]: <Required Confirmations Before Channel Open Number>
+      [remote_csv]: <Peer Unilateral Balance Output CSV Delay Number>
+      [remote_reserve]: <Minimum Tokens Peer Must Keep On Their Side Number>
+      [remote_max_htlcs]: <Maximum Slots For Attaching HTLCs Number>
+      [remote_max_pending_mtokens]: <Maximum HTLCs Value Millitokens String>
+      [remote_min_htlc_mtokens]: <Minimium HTLC Value Millitokens String>
+    }) -> {}
     capacity: <Capacity Tokens Number>
     chain: <Chain Id Hex String>
     commit_fee_tokens_per_vbyte: <Commitment Transaction Fee Number>
@@ -42,7 +52,9 @@ const weightPerVByte = 4;
     min_chain_output: <Minimum Chain Output Tokens Number>
     min_htlc_mtokens: <Minimum HTLC Millitokens String>
     partner_public_key: <Peer Public Key Hex String>
-    reject: <Reject Request Function>
+    reject: <Reject Request Function> ({
+      [reason]: <500 Character Limited Rejection Reason String>
+    }) -> {}
   }
 
   @event 'error'
@@ -129,10 +141,24 @@ module.exports = ({lnd}) => {
     const feeTok = Number(data.fee_per_kw) * weightPerVByte / weightPerKWeight;
     const id = data.pending_chan_id;
 
-    const resolveRequest = accept => sub.write({accept, pending_chan_id: id});
-
     return emitter.emit(channelRequestEvent, ({
-      accept: () => resolveRequest(true),
+      accept: (params) => {
+        if (!params) {
+          return sub.write({accept: true, pending_chan_id: id});
+        }
+
+        return sub.write({
+          accept: true,
+          csv_delay: params.remote_csv || undefined,
+          in_flight_max_msat: params.remote_max_pending_mtokens || undefined,
+          max_htlc_count: params.remote_max_htlcs || undefined,
+          min_accept_depth: params.min_confirmations || undefined,
+          min_htlc_in: params.remote_min_htlc_mtokens || undefined,
+          pending_chan_id: id,
+          reserve_sat: params.remote_reserve || undefined,
+          upfront_shutdown: params.cooperative_close_address || undefined,
+        });
+      },
       capacity: Number(data.funding_amt),
       chain: data.chain_hash.reverse().toString('hex'),
       commit_fee_tokens_per_vbyte: feeTok,
@@ -145,7 +171,17 @@ module.exports = ({lnd}) => {
       min_chain_output: Number(data.dust_limit),
       min_htlc_mtokens: data.min_htlc,
       partner_public_key: data.node_pubkey.toString('hex'),
-      reject: () => resolveRequest(false),
+      reject: (params) => {
+        if (!params) {
+          return sub.write({accept: false, pending_chan_id: id});
+        }
+
+        return sub.write({
+          accept: false,
+          error: params.reason || undefined,
+          pending_chan_id: id,
+        });
+      },
     }));
   });